+++ src/scene/shader-lib/chunks/gsplat/vert/gsplatColor.js
@@ -1,0 +1,9 @@
+export default /* glsl */`
+
+uniform mediump sampler2D splatColor;
+
+vec4 readColor(in SplatState state) {
+    return texelFetch(splatColor, state.uv, 0);
+}
+
+`;
