+++ src/scene/shader-lib/chunks/gsplat/vert/gsplatCompressedData.js
@@ -1,6 +1,4 @@
 export default /* glsl */`
-#include "gsplatUnpackVS"
-
 attribute vec3 vertex_position;         // xy: cornerUV, z: render order offset
 attribute uint vertex_id_attrib;        // render order base
 
@@ -16,6 +14,56 @@ vec4 chunkDataC;    // x: scale_min_z, y: scale_max_x, z: scale_max_y, w: scale_
 vec4 chunkDataD;    // x: min_r, y: min_g, z: min_b, w: max_r
 vec4 chunkDataE;    // x: max_g, y: max_b, z: unused, w: unused
 uvec4 packedData;   // x: position bits, y: rotation bits, z: scale bits, w: color bits
+
+vec3 unpack111011(uint bits) {
+    return vec3(
+        float(bits >> 21u) / 2047.0,
+        float((bits >> 11u) & 0x3ffu) / 1023.0,
+        float(bits & 0x7ffu) / 2047.0
+    );
+}
+
+vec4 unpack8888(uint bits) {
+    return vec4(
+        float(bits >> 24u) / 255.0,
+        float((bits >> 16u) & 0xffu) / 255.0,
+        float((bits >> 8u) & 0xffu) / 255.0,
+        float(bits & 0xffu) / 255.0
+    );
+}
+
+const float norm = 1.0 / (sqrt(2.0) * 0.5);
+
+vec4 unpackRotation(uint bits) {
+    float a = (float((bits >> 20u) & 0x3ffu) / 1023.0 - 0.5) * norm;
+    float b = (float((bits >> 10u) & 0x3ffu) / 1023.0 - 0.5) * norm;
+    float c = (float(bits & 0x3ffu) / 1023.0 - 0.5) * norm;
+    float m = sqrt(1.0 - (a * a + b * b + c * c));
+
+    uint mode = bits >> 30u;
+    if (mode == 0u) return vec4(m, a, b, c);
+    if (mode == 1u) return vec4(a, m, b, c);
+    if (mode == 2u) return vec4(a, b, m, c);
+    return vec4(a, b, c, m);
+}
+
+mat3 quatToMat3(vec4 R) {
+    float x = R.x;
+    float y = R.y;
+    float z = R.z;
+    float w = R.w;
+    return mat3(
+        1.0 - 2.0 * (z * z + w * w),
+              2.0 * (y * z + x * w),
+              2.0 * (y * w - x * z),
+              2.0 * (y * z - x * w),
+        1.0 - 2.0 * (y * y + w * w),
+              2.0 * (z * w + x * y),
+              2.0 * (y * w + x * z),
+              2.0 * (z * w - x * y),
+        1.0 - 2.0 * (y * y + z * z)
+    );
+}
 
 // calculate the current splat index and uvs
 bool readCenter(out SplatState state) {
@@ -81,56 +129,4 @@ void readCovariance(in SplatState state, out vec3 covA, out vec3 covB) {
     covA = vec3(dot(M[0], M[0]), dot(M[0], M[1]), dot(M[0], M[2]));
     covB = vec3(dot(M[1], M[1]), dot(M[1], M[2]), dot(M[2], M[2]));
 }
-
-// spherical Harmonics
-
-#if SH_BANDS > 0
-
-uniform highp usampler2D shTexture0;
-uniform highp usampler2D shTexture1;
-uniform highp usampler2D shTexture2;
-
-vec4 sunpack8888(in uint bits) {
-    return vec4((uvec4(bits) >> uvec4(0u, 8u, 16u, 24u)) & 0xffu) * (8.0 / 255.0) - 4.0;
-}
-
-void readSHData(in SplatState state, out vec3 sh[15]) {
-    // read the sh coefficients
-    uvec4 shData0 = texelFetch(shTexture0, state.uv, 0);
-    uvec4 shData1 = texelFetch(shTexture1, state.uv, 0);
-    uvec4 shData2 = texelFetch(shTexture2, state.uv, 0);
-
-    vec4 r0 = sunpack8888(shData0.x);
-    vec4 r1 = sunpack8888(shData0.y);
-    vec4 r2 = sunpack8888(shData0.z);
-    vec4 r3 = sunpack8888(shData0.w);
-
-    vec4 g0 = sunpack8888(shData1.x);
-    vec4 g1 = sunpack8888(shData1.y);
-    vec4 g2 = sunpack8888(shData1.z);
-    vec4 g3 = sunpack8888(shData1.w);
-
-    vec4 b0 = sunpack8888(shData2.x);
-    vec4 b1 = sunpack8888(shData2.y);
-    vec4 b2 = sunpack8888(shData2.z);
-    vec4 b3 = sunpack8888(shData2.w);
-
-    sh[0] =  vec3(r0.x, g0.x, b0.x);
-    sh[1] =  vec3(r0.y, g0.y, b0.y);
-    sh[2] =  vec3(r0.z, g0.z, b0.z);
-    sh[3] =  vec3(r0.w, g0.w, b0.w);
-    sh[4] =  vec3(r1.x, g1.x, b1.x);
-    sh[5] =  vec3(r1.y, g1.y, b1.y);
-    sh[6] =  vec3(r1.z, g1.z, b1.z);
-    sh[7] =  vec3(r1.w, g1.w, b1.w);
-    sh[8] =  vec3(r2.x, g2.x, b2.x);
-    sh[9] =  vec3(r2.y, g2.y, b2.y);
-    sh[10] = vec3(r2.z, g2.z, b2.z);
-    sh[11] = vec3(r2.w, g2.w, b2.w);
-    sh[12] = vec3(r3.x, g3.x, b3.x);
-    sh[13] = vec3(r3.y, g3.y, b3.y);
-    sh[14] = vec3(r3.z, g3.z, b3.z);
-}
-
-#endif
 `;
